pyproject.toml
[project]
name = "gaia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "affine==2.4.0",
    "aioboto3==14.3.0",
    "aiobotocore==2.22.0",
    "aiofiles==24.1.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.10.11",
    "aioitertools==0.12.0",
    "aiosignal==1.3.2",
    "alembic==1.16.1",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asciitree==0.3.3",
    "async-substrate-interface==1.1.1",
    "async-timeout==5.0.1",
    "asyncpg==0.30.0",
    "asyncstdlib==3.13.1",
    "attrs==25.3.0",
    "azure-core==1.35.0",
    "azure-storage-blob==12.26.0",
    "base58==2.1.1",
    "bittensor-commit-reveal==0.1.0",
    "bittensor-wallet==3.0.10",
    "blosc==1.11.3",
    "boto3==1.37.3",
    "botocore==1.37.3",
    "bt-decode==0.6.0",
    "cachetools==5.5.2",
    "cchardet==2.1.7 ; python_full_version < '3.12'",
    "cdsapi==0.7.6",
    "certifi==2025.8.3",
    "cffi==1.17.1",
    "cfgrib==0.9.15.0",
    "cftime==1.6.4.post1",
    "charset-normalizer==3.4.3",
    "click==8.2.0",
    "click-plugins==1.1.1",
    "cligj==0.7.2",
    "cloudpickle==3.1.1",
    "cmdstanpy==1.2.5",
    "colorama==0.4.6",
    "contourpy==1.3.2",
    "cryptography>=43.0.1,<43.1",
    "cycler==0.12.1",
    "cytoolz==1.0.1",
    "dask==2025.5.1",
    "decorator==5.2.1",
    "eccodes==2.42.0",
    "ecdsa==0.19.1",
    "ecmwf-datastores-client==0.1.0",
    "einops==0.8.1",
    "eth-hash==0.7.1",
    "eth-keys==0.7.0",
    "eth-typing==5.0.0",
    "eth-utils==2.3.2",
    "exceptiongroup==1.3.0",
    "fastapi==0.112.0",
    "fasteners==0.19",
    "fiber",
    "filelock==3.19.1",
    "findlibs==0.1.1",
    "flexcache==0.3",
    "flexparser==0.4",
    "fonttools==4.58.0",
    "frozenlist==1.6.0",
    "fsspec==2025.3.2",
    "gcsfs==2025.3.2",
    "google-api-core==2.25.0rc1",
    "google-auth==2.40.2",
    "google-auth-oauthlib==1.2.2",
    "google-cloud-core==2.4.3",
    "google-cloud-storage==3.1.0",
    "google-crc32c==1.7.1",
    "google-resumable-media==2.7.2",
    "googleapis-common-protos==1.70.0",
    "greenlet==3.2.2",
    "h11==0.16.0",
    "h3==4.2.2",
    "h5netcdf==1.6.1",
    "h5py==3.13.0",
    "hf-xet==1.1.8",
    "holidays==0.72",
    "httpcore==1.0.9",
    "httpx==0.27.2",
    "huggingface-hub==0.34.4",
    "idna==3.10",
    "imageio==2.37.0",
    "importlib-metadata==8.7.0",
    "importlib-resources==6.5.2",
    "intake==2.0.8",
    "intake-xarray==2.0.0",
    "isodate==0.7.2",
    "jinja2==3.1.6",
    "jmespath==1.0.1",
    "kerchunk==0.2.7",
    "kiwisolver==1.4.8",
    "lazy-loader==0.4",
    "lightning-utilities==0.14.3",
    "linkify-it-py==2.0.3",
    "locket==1.0.0",
    "loguru==0.7.3",
    "lz4==4.4.4",
    "mako==1.3.10",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "matplotlib==3.10.3",
    "mdit-py-plugins==0.4.2",
    "mdurl==0.1.2",
    "memray==1.17.2",
    "metpy==1.7.0",
    "microsoft-aurora==1.6.0",
    "more-itertools==10.7.0",
    "mpmath==1.3.0",
    "msgpack==1.1.1",
    "multidict==6.4.3",
    "multiurl==0.3.5",
    "netaddr==1.3.0",
    "netcdf4==1.7.2",
    "networkx==3.4.2",
    "numcodecs==0.13.1",
    "numpy==2.2.6",
    "nvidia-cublas-cu12==12.8.4.1",
    "nvidia-cuda-cupti-cu12==12.8.90",
    "nvidia-cuda-nvrtc-cu12==12.8.93",
    "nvidia-cuda-runtime-cu12==12.8.90",
    "nvidia-cudnn-cu12==9.10.2.21",
    "nvidia-cufft-cu12==11.3.3.83",
    "nvidia-cufile-cu12==1.13.1.3",
    "nvidia-curand-cu12==10.3.9.90",
    "nvidia-cusolver-cu12==11.7.3.90",
    "nvidia-cusparse-cu12==12.5.8.93",
    "nvidia-cusparselt-cu12==0.7.1",
    "nvidia-nccl-cu12==2.27.3",
    "nvidia-nvjitlink-cu12==12.8.93",
    "nvidia-nvtx-cu12==12.8.90",
    "oauthlib==3.2.2",
    "orjson==3.10.18",
    "packaging==25.0",
    "pandas==2.3.2",
    "partd==1.4.2",
    "password-strength==0.0.3.post2",
    "patsy==1.0.1",
    "pillow==11.3.0",
    "pint==0.24.4",
    "pip==25.2",
    "platformdirs==4.3.8",
    "pooch==1.8.2",
    "propcache==0.3.1",
    "properscoring==0.1",
    "prophet==1.1.6",
    "proto-plus==1.26.1",
    "protobuf==6.31.0",
    "psutil==7.0.0",
    "psycopg2-binary==2.9.10",
    "py-bip39-bindings==0.1.11",
    "py-ed25519-zebra-bindings==1.2.0",
    "py-sr25519-bindings==0.2.2",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pycryptodome==3.22.0",
    "pydantic>=2.0,<2.10",
    "pygments==2.19.1",
    "pyinstrument==5.0.2",
    "pyjwt==2.10.1",
    "pynacl==1.5.0",
    "pyparsing==3.2.3",
    "pyproj==3.7.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "pytorch-lightning==2.5.1.post0",
    "pytorch-msssim==1.0.0",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "rasterio==1.4.3",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "rich==14.0.0",
    "rsa==4.9.1",
    "s3transfer==0.11.3",
    "safetensors==0.6.2",
    "scalecodec==1.2.11",
    "scikit-image==0.25.2",
    "scipy==1.15.3",
    "setuptools==80.9.0",
    "shapely==2.1.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.40",
    "stanio==0.5.1",
    "starlette==0.37.2",
    "statsmodels==0.14.4",
    "substrate-interface==1.7.10",
    "sympy==1.14.0",
    "tenacity==9.0.0",
    "textual==3.4.0",
    "tifffile==2025.5.10",
    "timm==1.0.19",
    "toml==0.10.0",
    "tomli==2.2.1",
    "toolz==1.0.0",
    "torch==2.8.0",
    "torchmetrics==1.7.1",
    "torchvision==0.23.0",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "triton==3.4.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "uc-micro-py==1.0.3",
    "ujson==5.10.0",
    "urllib3==2.5.0",
    "uvicorn==0.30.5",
    "uvloop==0.21.0",
    "websocket-client==1.8.0",
    "websockets==15.0.1",
    "wheel==0.45.1",
    "wrapt==1.17.2",
    "xarray==2025.6.1",
    "xhistogram==0.3.2",
    "xskillscore==0.0.26",
    "xxhash==3.5.0",
    "yarl==1.20.0",
    "zarr==2.18.3",
    "zipp==3.21.0",
]

[project.optional-dependencies]
dev = [
    "isort>=6.0.1",
]
miner = [
    "aioboto3==14.3.0",
    "boto3==1.37.3",
    "botocore==1.37.3",
    "s3transfer==0.11.3",
    "gcsfs==2025.3.2",
    "google-api-core==2.25.0rc1",
    "google-auth==2.40.2",
    "google-auth-oauthlib==1.2.2",
    "google-cloud-core==2.4.3",
    "google-cloud-storage==3.1.0",
    "google-crc32c==1.7.1",
    "google-resumable-media==2.7.2",
    "googleapis-common-protos==1.70.0",
    "azure-core==1.35.0",
    "azure-storage-blob==12.26.0",
]
validator = [
    "alembic==1.16.1",
    "sqlalchemy==2.0.40",
    "asyncpg==0.30.0",
    "psycopg2-binary==2.9.10",
]

[tool.uv.sources]
fiber = { git = "https://github.com/rayonlabs/fiber.git", rev = "0ba1950fb7f8857265fd3854c727a4b3d1a83452" }
